feat: altera script antes da instalacao

+++ scripts/before-install.sh
@@ -1,2 +1,5 @@
 #!/bin/sh
-rm /home/ec2-user/*.jar
+JAR=$(find /home/ec2-user/ -iname '*.jar' -type -f)
+if [ $JAR ]; then
+  rm /home/ec2-user/*.jar
+fi
